+++ packages/safe-core-sdk/tests/threshold.test.ts
@@ -2,8 +2,9 @@ import chai from 'chai'
 import chaiAsPromised from 'chai-as-promised'
 import { deployments, ethers, waffle } from 'hardhat'
 import EthersSafe, { ContractNetworksConfig, EthersAdapter } from '../src'
-import { getAccounts } from './utils/setupConfig'
 import { getMultiSend, getSafeWithOwners } from './utils/setupContracts'
+import { getAccounts } from './utils/setupTestNetwork'
+
 chai.use(chaiAsPromised)
 
 describe('Safe Threshold', () => {
